requirements.txt
langchain-groq
langgraph
streamlit
langchain
langchain-core
langchain-community
duckduckgo-search
python-dotenv
requests
langgraph-checkpoint-sqlite
ddgs
